Enable packaging of the repo
These changes enable pathlesstaken to be packaged. The code layout is
improved and various other changes are made as well. Imports are
fixed for the different ways the code may be run. A repository level
main can be run if calling this code from the git repo, vs. PyPi.
