+++ .gitignore
@@ -3,3 +3,6 @@ py2
 py3
 venv
 .tox
+build/
+dist/
+*.egg-info/
