+++ src/pathlesstaken/version.py
@@ -1,0 +1,3 @@
+"""Version module for pathlesstaken."""
+
+__version__ = "v1.0.2"
